Add script to change current username
This commit adds a new script, `script.bat`, that allows users to change their current username. The script prompts the user to enter a new username and uses the `net user` command to change the username. It also displays a success message if the username change is successful, or an error message if it fails.